moved setup wrapper script stuff out of mrjob.runner
